Refactor 상세 게시글 페이지 댓글 섹션 분리
분량이 너무 많아 분리합니다.
