+++ src/pages/postdetail/PostDetail.module.css
@@ -117,50 +117,6 @@
   background-color: #e5b3b3 !important;
 }
 
-.commentForm {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-top: 20px;
-  border-top: 1px solid #00000040;
-}
-
-.commentInput {
-  width: 100%;
-  min-height: 120px;
-  padding: 10px;
-  border: 1px solid #00000040;
-  border-radius: 5px;
-  resize: none;
-  font-size: 14px;
-}
-
-.commentSubmitButtonWrapper {
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-}
-
-.commentSubmit {
-  margin-top: 10px;
-  padding: 10px 35px;
-  border: none;
-  border-radius: 15px;
-  background-color: #aca0eb;
-  color: #ffffff;
-  font-weight: bold;
-}
-
-.commentList {
-  display: flex;
-  flex-direction: column;
-  row-gap: 20px; /* 댓글 간 간격 유지 */
-  width: 100%;
-  max-width: 100%; /* 화면 초과 방지 */
-}
-
 @media (max-width: 640px) {
   .postDetail {
     max-width: 100%; /* 전체 화면 너비에 맞게 확장 */
@@ -209,23 +165,5 @@
   .btnReaction {
     width: 100px; /* 버튼 너비 축소 */
     height: 70px;
-  }
-
-  .commentForm {
-    padding-top: 10px; /* 내부 패딩 축소 */
-  }
-
-  .commentInput {
-    min-height: 100px; /* 최소 높이 축소 */
-    padding: 0; /* 내부 패딩 축소 */
-    width: 98%;
-  }
-
-  .commentSubmit {
-    padding: 8px 20px; /* 버튼 크기 축소 */
-  }
-
-  .commentList {
-    row-gap: 15px; /* 댓글 간 간격 축소 */
   }
 }
